feat: Added default values for configuration, tests & some refactoring.
- Added a new value `RESULTS_FILE_LIMIT` for configuration.
- Added optional values when reading from the config for default values instead of crashes if a key is incorrect.
- Added tests for config.
- Added a table of default config values to `README.md`.
- `getString(...)` to get config value is renamed to `getProperty(...)`.
- Some refactoring and other minor changes.

+++ src/test/java/CodeCheckTest.java
@@ -1,6 +1,8 @@
 import CodeCheck.CodeCheck;
 import CodeCheck.Log;
 import CodeCheck.Util;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -10,34 +12,60 @@ import java.util.List;
 import java.util.Scanner;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class CodeCheckTest {
 
-    private final String userDirectory = System.getProperty("user.dir");
-    private final String testConfPath = userDirectory + File.separator + "test-config.properties";
+    private static final String userDirectory = System.getProperty("user.dir");
+    private static final String testConfPath = userDirectory + File.separator + "test-config.properties";
     private final String resultFolder = "test-results";
     private final String resultNamePrefix = "test-result_{nr}";
     private final String resultFileName = "test-result_0.txt";
 
+
+    @BeforeEach
+    public void BeforeEach() {
+        delete(userDirectory + File.separator + resultFolder);
+    }
+
+    @AfterAll
+    public static void AfterAll() {
+        emptyTestConf();
+    }
+
     @Test
     public void executeTest() throws Exception {
-        delete(userDirectory + File.separator + resultFolder);
-
         writeConf();
         CodeCheck.execute();
 
         assertEquals(2, readResult().split("\n").length - 1);
+    }
 
-        emptyTestConf();
+    @Test
+    public void executeMinimalTest() throws Exception {
+        writeMinimalConf();
+        CodeCheck.execute();
+
+        assertEquals(2, readResult().split("\n").length - 1);
+    }
+
+    @Test
+    public void executeEmptyTest() {
+        writeEmptyConf();
+
+        Exception e = assertThrows(NullPointerException.class, CodeCheck::execute);
+
+        assertEquals(e.getMessage(),
+                "Cannot invoke \"Object.toString()\" because the return value of \"java.util.Properties.get(Object)\" is null");
     }
 
     private void delete(String path) {
         File dir = new File(path);
-        if (dir.isDirectory()) {
+        if (dir.isDirectory() && dir.getName().contains("test")) {
             Arrays.stream(dir.listFiles()).forEach(File::delete);
             dir.delete();
         } else {
-            Log.warning("Could not delete '%s' directory as it doesn't exist.".formatted(resultFolder));
+            Log.log("Could not delete '%s' directory as it doesn't exist.".formatted(resultFolder));
         }
     }
 
@@ -56,7 +84,27 @@ public class CodeCheckTest {
         conf.forEach(line -> Util.write(testConfPath, line));
     }
 
-    private void emptyTestConf() {
+    private void writeMinimalConf() {
+        List<String> conf = List.of(
+                "PATH_TO_RESULTS=" + resultFolder, // Is optional - used for testing.
+                "RESULT_NAME_PREFIX=" + resultNamePrefix, // Is optional - used for testing.
+                "PATH_TO_CODE=src/test/resources/"); // Is required.
+
+        Util.write(testConfPath, "# properties used under testing - this file will be overwritten during testing", false);
+        conf.forEach(line -> Util.write(testConfPath, line));
+    }
+
+    private void writeEmptyConf() {
+        List<String> conf = List.of(
+                "PATH_TO_RESULTS=" + resultFolder, // Is optional - used for testing.
+                "RESULT_NAME_PREFIX=" + resultNamePrefix // Is optional - used for testing.
+        );
+
+        Util.write(testConfPath, "# properties used under testing - this file will be overwritten during testing", false);
+        conf.forEach(line -> Util.write(testConfPath, line));
+    }
+
+    private static void emptyTestConf() {
         Util.write(testConfPath, "# empty -- properties used under testing -- this file will be overwritten during testing", false);
     }
 
@@ -80,9 +128,7 @@ public class CodeCheckTest {
             myReader.close();
 
         } catch (FileNotFoundException e) {
-            String errorMessage = "Testing result file not found. %s".formatted(e.getMessage());
-            Log.error(errorMessage);
-            throw new FileNotFoundException(errorMessage + "\n" + Arrays.toString(e.getStackTrace()));
+            throw new FileNotFoundException("Testing result file not found. %s\n%s".formatted(e.getMessage(), Arrays.toString(e.getStackTrace())));
         }
 
         return content.toString();
